Newly committed changes
Kikomeko Bashir and Lillian Nabisubi will be working on the User Interfaces of the web pages.
Kirumira Benjamin will be working on the user and administrator authentication.
Bahebwa Rashidah  will be dealing with the forms and lastly,Luswata Andrew will be working on the database management i.e. the displaying of information in the database on the pages and adding information input in the pages into the database.